+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-14T08:35:33.053Z"
+    created: Optional[str] = "2026-07-14T09:44:44.987Z"
 
-    updated: Optional[str] = "2026-07-14T08:35:33.053Z"
+    updated: Optional[str] = "2026-07-14T09:44:44.987Z"
 
     updated_by_id: OptionalNullable[str] = UNSET