+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXFW9KAFSMFQGVRWNP9W5X6G"
+        "01KXG07WKZ0GAGPFJMYJP1YCYW"
     )
     r"""The unique identifier of the data source"""