+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXFW9KAJCPT5C9CBN1C51V1T"
+        "01KXG07WM0N458731FZZQ03D26"
     )
     r"""The unique identifier of the data source"""